conduit_raw/requirements.txt
lmdb
bitcoinx
requests>=2.21.0
dnspython~=1.16.0
mysqlclient
pyzmq
cbor2
aiohttp
fasteners
pygments
